Add copyright, fix documentation
