+++ java/streams/src/main/java/org/dbsp/algebraic/IntegerGroup.java
@@ -21,37 +21,29 @@
  * SOFTWARE.
  */
 
-package org.dbsp.formal;
-
-import org.dbsp.algebraic.TimeFactory;
-import org.dbsp.types.IStream;
-import org.dbsp.types.StreamType;
-import org.dbsp.types.Type;
-
-import java.util.function.Function;
+package org.dbsp.algebraic;
 
 /**
- * An operator that works on streams.  It delays the input stream by 1 clock.
- * @param <T> Concrete implementation of elements of the input and output stream.
+ * The group of integer values.
+ * This implementation uses finite Java integers.
  */
-public class DelayOperator<T, F extends TimeFactory> extends
-        UniformUnaryOperator<IStream<T>> {
-    final Type<T> elementType;
+public class IntegerGroup implements Group<Integer> {
+    private IntegerGroup() {}
 
-    public DelayOperator(Type<T> type, F factory) {
-        super(new StreamType<T>(type, factory));
-        this.elementType = type;
-        if (!type.isStream())
-            throw new RuntimeException("Delay must be applied to a stream type, not to " + type);
+    public static final IntegerGroup instance = new IntegerGroup();
+
+    @Override
+    public Integer minus(Integer data) {
+        return -data;
     }
 
     @Override
-    public Function<IStream<T>, IStream<T>> getComputation() {
-        return (IStream<T> s) -> s.delay(this.elementType.getGroup());
+    public Integer add(Integer left, Integer right) {
+        return left + right;
     }
 
     @Override
-    public String toString() {
-        return "z";
+    public Integer zero() {
+        return 0;
     }
 }
