+++ java/streams/src/main/java/org/dbsp/algebraic/Monoid.java
@@ -1,6 +1,44 @@
+/*
+ * Copyright (c) 2021 VMware, Inc.
+ * SPDX-License-Identifier: MIT
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
 package org.dbsp.algebraic;
 
+/**
+ * A monoid is an algebraic structure with an associative operation and a
+ * neutral element called zero.  All our monoids are commutative too.
+ * @param <T>  Java representation of the Monoid values.
+ */
 public interface Monoid<T> {
+    /**
+     * Associative function that adds two values in the monoid.
+     * @param left   Left value to add.
+     * @param right  Right value to add.
+     * @return       The sum of left and right.
+     */
     T add(T left, T right);
+
+    /**
+     * The zero element of the monoid.
+     */
     T zero();
 }
