+++ java/streams/src/main/java/org/dbsp/formal/UnaryOperator.java
@@ -1,3 +1,26 @@
+/*
+ * Copyright (c) 2021 VMware, Inc.
+ * SPDX-License-Identifier: MIT
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
 package org.dbsp.formal;
 
 import org.dbsp.algebraic.TimeFactory;
@@ -44,7 +67,7 @@ public abstract class UnaryOperator<T, S> {
     /**
      * @return A lifted version of this operator.
      */
-    public <F extends TimeFactory> UnaryOperator<IStream<T>, IStream<S>> lift(F factory) {
+    public UnaryOperator<IStream<T>, IStream<S>> lift(TimeFactory factory) {
         return new UnaryOperator<IStream<T>, IStream<S>>(
                 new StreamType<T>(this.inputType, factory),
                 new StreamType<S>(this.outputType, factory)) {
